modifed journal.js component
